fix: alert - provider as jotai atom
