--- apps/web/src/components/provider/Alert/AlertProvider/index.ts
@@ -1,1 +1,0 @@
-export { AlertProvider as default } from './AlertProvider';
