+++ apps/web/src/components/provider/Alert/component/AlertModal.tsx
@@ -1,0 +1,69 @@
+import type { AlertData } from '../atom';
+import { Button, Modal, ModalBody, ModalContent, ModalFooter, ModalHeader } from '@heroui/react'
+import { useEffect, useState } from 'react'
+import { useTranslation } from 'react-i18next'
+import StateRender from '@utils/StateRender';
+import useAlert from '../hooks/useAlert';
+
+export const AlertModal = () => {
+  const { t } = useTranslation('alert')
+  const { alertList, setAlertList } = useAlert()
+  const [displayMessage, setDisplayMessage] = useState<AlertData>()
+
+  useEffect(() => {
+    if (!displayMessage && alertList.length) {
+      const [tempMessage, ...alertMessageOther] = alertList
+      setAlertList(alertMessageOther)
+      setDisplayMessage(tempMessage)
+    }
+  }, [displayMessage, alertList, setAlertList])
+
+  function alertDisplayChange() {
+    if (displayMessage?.important) return
+    if (displayMessage?.resolver) {
+      displayMessage.resolver(false)
+    }
+    setDisplayMessage(undefined)
+  }
+  function cancelAction() {
+    if (!displayMessage?.resolver) return
+    displayMessage.resolver(false)
+    setDisplayMessage(undefined)
+  }
+  function confirmAction() {
+    if (!displayMessage?.resolver) return
+    displayMessage.resolver(true)
+    setDisplayMessage(undefined)
+  }
+
+  return <Modal isOpen={Boolean(displayMessage)} onOpenChange={alertDisplayChange}>
+    <ModalContent>
+      {() => <>
+        <ModalHeader>{t('modal.title')}</ModalHeader>
+        <ModalBody>
+          <div>
+            {displayMessage?.message}
+          </div>
+        </ModalBody>
+        <ModalFooter>
+          <StateRender.Boolean
+            state={Boolean(displayMessage?.cancel)}
+            render={{
+              true: <Button className={displayMessage?.cancel?.colorClass} onPress={cancelAction}>
+                {displayMessage?.cancel?.message}
+              </Button>
+            }}
+          />
+          <StateRender.Boolean
+            state={Boolean(displayMessage?.confirm)}
+            render={{
+              true: <Button className={displayMessage?.confirm?.colorClass} onPress={confirmAction}>
+                {displayMessage?.confirm?.message}
+              </Button>
+            }}
+          />
+        </ModalFooter>
+      </>}
+    </ModalContent>
+  </Modal>
+}
