+++ apps/web/src/pages/:selectDate/schedule/create.tsx
@@ -19,7 +19,7 @@ export default function CreateSchedule() {
   const { selectDate } = useParams()
   const [queryParams] = useSearchParams(defaultSearchParams)
   const [year, month, date] = selectDate ? selectDate.split('-').map(v => +v) : [0, 0, 0]
-  const { showAlert } = useAlert()
+  const { pushAlert } = useAlert()
 
   const createScheduleWithExercisePlans = useCreateScheduleWithExercisePlans()
   async function startFitnessTime(exerciseIdxList: number[]) {
@@ -30,7 +30,7 @@ export default function CreateSchedule() {
       if (result?.id) {
         gotoScheduleDetail(result?.id, `${year}-${month}-${date}`, { replace: true })
       } else {
-        showAlert({
+        pushAlert({
           message: 'Error... Something Wrong!'
         })
       }
